# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RescueBase API",
    "description": "API for Sanitätslager, Rucksackchecks, Chargen and Nachfüllaufträge.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Reverse-proxied API base path"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "rb_session"
      }
    },
    "schemas": {
      "MpgInput": {
        "type": "object",
        "additionalProperties": true
      },
      "MpgRecord": {
        "type": "object",
        "additionalProperties": true
      },
...
    }
  },
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/mpg/models": {
      "get": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_modelsList",
        "responses": {
          "200": {
            "description": "MPG models",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MpgRecord"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_modelsCreate",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MpgInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Model created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpgRecord"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/models/{id}": {
      "patch": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_modelsUpdate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MpgInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Model updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpgRecord"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/models/{id}/requirements": {
      "post": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_requirement",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MpgInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Requirement created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpgRecord"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/models/{id}/review": {
      "post": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_modelReview",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MpgInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Requirements reviewed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpgRecord"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/devices": {
      "get": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_devicesList",
        "responses": {
          "200": {
            "description": "Devices",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MpgRecord"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_deviceCreate",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MpgInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Device created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpgRecord"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/devices/{id}": {
      "get": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_device",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Device record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpgRecord"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_deviceUpdate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MpgInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Device updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpgRecord"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/devices/{id}/release": {
      "post": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_deviceRelease",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MpgInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Device released",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpgRecord"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/devices/{id}/retire": {
      "post": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_deviceRetire",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MpgInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Device retired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpgRecord"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/devices/{id}/inspections": {
      "post": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_inspectionCreate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MpgInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Inspection draft",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpgRecord"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/inspections/{id}": {
      "patch": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_inspectionUpdate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MpgInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inspection updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpgRecord"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/inspections/{id}/finalize": {
      "post": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_inspectionFinalize",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MpgInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Inspection finalized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpgRecord"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/inspections/{id}/correct": {
      "post": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_inspectionCorrect",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MpgInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Correction created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpgRecord"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/people": {
      "get": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_peopleList",
        "responses": {
          "200": {
            "description": "People",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MpgRecord"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_personCreate",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MpgInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Person created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpgRecord"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/people/{id}": {
      "patch": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_personUpdate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MpgInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Person updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpgRecord"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/trainings": {
      "get": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_trainingList",
        "responses": {
          "200": {
            "description": "Trainings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MpgRecord"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_trainingCreate",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MpgInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Training created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpgRecord"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/trainings/{id}": {
      "get": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_training",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Training",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpgRecord"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/trainings/{id}/confirm": {
      "post": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_trainingConfirm",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MpgInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Confirmation stored",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpgRecord"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/trainings/{id}/finalize": {
      "post": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_trainingFinalize",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MpgInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Training finalized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpgRecord"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/devices/{id}/glucose-controls": {
      "post": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_glucose",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MpgInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Control finalized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpgRecord"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/glucose-controls/{id}/resolve": {
      "post": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_glucoseResolve",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MpgInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Control resolved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpgRecord"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/devices/{id}/incidents": {
      "post": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_incident",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MpgInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Incident created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpgRecord"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/incidents/{id}": {
      "patch": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_incidentUpdate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MpgInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Incident updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpgRecord"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/cylinders": {
      "get": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_cylindersList",
        "responses": {
          "200": {
            "description": "Cylinders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MpgRecord"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_cylinderCreate",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MpgInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cylinder created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpgRecord"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/cylinders/{id}": {
      "patch": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_cylinderUpdate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MpgInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cylinder updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpgRecord"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/cylinders/{id}/assign": {
      "post": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_cylinderAssign",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MpgInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cylinder assigned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpgRecord"
                }
              }
            }
          }
        }
      }
    },
    "/mpg/cylinders/{id}/return": {
      "post": {
        "tags": [
          "MPG"
        ],
        "operationId": "MpgController_cylinderReturn",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MpgInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cylinder returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpgRecord"
                }
              }
            }
          }
        }
      }
    },
...
  }
}